refactor in .py predict_POOL_handle.py
